moved helpers functions to imported helper file
